Atualização Restaurante #19

+++ codigo/comidinhasVeg/src/main/java/javaquinho/comidinhas/models/Mesa.java
@@ -1,10 +1,14 @@
 package javaquinho.comidinhas.models;
 
+import com.fasterxml.jackson.annotation.JsonBackReference;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -27,7 +31,7 @@ public class Mesa {
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", unique = true, nullable = false)
-    private int idMesa;
+    private Long idMesa;
 
 	@Column(name = "capacidade", nullable = false)
 	private int capacidade;
@@ -64,7 +68,7 @@ public class Mesa {
 		return (quantPessoas <= capacidade && !ocupada);
 	}
 
-	public int getIdMesa(){
+	public Long getIdMesa(){
 		return idMesa;
 	}
